requirements.txt
# Tested with Python 3.14.x; LangChain may emit Pydantic v1 notices — silenced in warn_filters.py.
streamlit>=1.28.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
pypdf>=5.0.0
python-dotenv>=1.0.0
